Separate test files slightly
